refactor: add frendly names for functions
